requirements.txt
streamlit>=1.46.1
pandas
pydantic
openai>=1.96.1
google-genai>=1.25.0
langchain>=0.3.26
langchain-community>=0.3.27
langchain-openai>=0.3.28 